clippy: finish the test-suite and example lint tail
Close the serially-revealed remainder of the -D warnings gate:

- approx_constant: use f64::consts::TAU in the rand Box-Muller host
  oracle; move arbitrary 3.14 test floats off the PI pattern
- needless_range_loop: iterator/enumerate form in seven test loops and
  bench_mandelbrot's CPU reference
- single_match: gpu_queue's must-not-panic probes drop the empty match
- useless_vec, unnecessary_cast (icb_metal), overindented doc list item
- dead_code / unused_assignments: allow-with-reason on the
  #[quanta::device] fns that only the wasm-route registry consumes and
  on the intentionally dead local init that wasm_local_renaming
  exercises across the if/else merge
- keep the {OpCase, cases} re-export under an allow: the op_matrix test
  binary consumes it while the differential binary compiles the same
  shared module without touching it
- assertions_on_constants: op_matrix's macOS guard becomes an explicit
  cfg! + panic

rustfmt drift in the vulkan driver files swept along.

+++ tests/icb_metal.rs
@@ -48,9 +48,9 @@ fn icb_metal_record_two_dispatches_matches_direct() {
         }
     };
     wave_a.bind(0, &field_a);
-    let mut p1 = gpu.dispatch(&wave_a, count as u32).unwrap();
+    let mut p1 = gpu.dispatch(&wave_a, count).unwrap();
     p1.wait().unwrap();
-    let mut p2 = gpu.dispatch(&wave_a, count as u32).unwrap();
+    let mut p2 = gpu.dispatch(&wave_a, count).unwrap();
     p2.wait().unwrap();
     let direct = field_a.read().unwrap();
 
@@ -66,7 +66,7 @@ fn icb_metal_record_two_dispatches_matches_direct() {
             return;
         }
     };
-    let groups = [count as u32 / wave_b.workgroup_size[0], 1, 1];
+    let groups = [count / wave_b.workgroup_size[0], 1, 1];
     if let Err(e) = icb.record_dispatch(&wave_b, groups) {
         eprintln!("skipping: ICB record not supported: {e}");
         return;
